warehouse/requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
ipython==8.12.0
requests==2.28.2

